FIX: minor fixes and refactoring
